Unify hvl_replay.h to use one-common source, to minimize risk of forgetting to update all the different implementations.
(Current state: some constifications were missing, and the loading APIs are not 100% uniform)
